Change the meaning of 'factory' and add tests
The name 'factory' used to denote the type of the ProviderInterface#get
method, and the InVessel#factory method was a shortcut for instancing a
DefaultProvider, assigning its 'get' method and registering it.

A 'factory' now means a function that corresponds to a call to the 'get'
method on a container with a fixed key. The InVessel#factory method was
changed to create and return such kind of function, and no longer
registers anything.

The term 'dependency-getter' was created to represent anything that can
be passed to InVessel#provider, i.e. a provider or a 'dependency-getter
function'. The latter term carries the former meaning of 'factory'.

Tests for 'factory' were added.